requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-openai
langchain-google-genai
langchain_experimental
beautifulsoup4
faiss-cpu
chromadb
scrapetube
pydub
yt_dlp
ffmpeg
black
isort
ffmpeg-python


